requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg 
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
strawberry-graphql[fastapi]